Refactor frame resource management and abstract away boilerplate code for beginning and end of frame.
Frames resources are no longer associated with the swapchain image index. Instead, we use a simple frame number index that simply increments / wraps each frame to index frame resources. This simplifies resource management and allows the CPU to run ahead a specific number of frames, independent of the number of swapchain images.